requirements.txt
numpy==2.0.2
packaging==26.0
pdf2image==1.17.0
pillow==11.3.0
pytesseract==0.3.13
